+++ src/Ext/RadSite/Body.h
@@ -22,12 +22,14 @@ public:
 	class ExtData final : public Extension<RadSiteClass>
 	{
 	public:
+		int CreationFrame;
 		WeaponTypeClass* Weapon;
 		RadTypeClass* Type;
 		HouseClass* RadHouse;
 		TechnoClass* RadInvoker;
 
 		ExtData(RadSiteClass* OwnerObject) : Extension<RadSiteClass>(OwnerObject)
+			, CreationFrame { 0 }
 			, RadHouse { nullptr }
 			, RadInvoker { nullptr }
 			, Type {}